+++ src/internal/extract/extract.go
@@ -1,11 +1,11 @@
-package common
+package extract
 
 import (
 	"math/rand"
 	"time"
 )
 
-func SelectRandomString(s []string) string {
+func RandomString(s []string) string {
 	rand.NewSource(time.Now().UnixNano())
 	selectedString := s[rand.Intn(len(s))]
 	return selectedString
